+++ public/components/toolbar/spatial_filter/draw_tooltip.tsx
@@ -7,7 +7,7 @@ import maplibregl, { Map as Maplibre, MapEventType, Popup } from 'maplibre-gl';
 import React, { Fragment, useEffect, useRef } from 'react';
 import { i18n } from '@osd/i18n';
 import { FILTER_DRAW_MODE } from '../../../../common';
-import {isEscapeKey} from "../../../../common/util";
+import { isEscapeKey } from '../../../../common/util';
 
 interface Props {
   map: Maplibre;
@@ -17,20 +17,24 @@ interface Props {
 
 const X_AXIS_GAP_BETWEEN_CURSOR_AND_POPUP = -12;
 const Y_AXIS_GAP_BETWEEN_CURSOR_AND_POPUP = 0;
+const KEY_UP_EVENT_TYPE = 'keyup';
+const MOUSE_MOVE_EVENT_TYPE = 'mousemove';
+const MOUSE_OUT_EVENT_TYPE = 'mouseout';
 
 const getTooltipContent = (mode: FILTER_DRAW_MODE): string => {
   switch (mode) {
     case FILTER_DRAW_MODE.POLYGON:
       return i18n.translate('maps.drawFilterPolygon.tooltipContent', {
-        defaultMessage: 'Click to start shape. Click for vertex. Double click to finish.',
+        defaultMessage:
+          'Click to start shape. Click for vertex. Double click to finish, [esc] to cancel',
       });
     case FILTER_DRAW_MODE.RECTANGLE:
       return i18n.translate('maps.drawFilterRectangle.tooltipContent', {
-        defaultMessage: 'Click and drag to draw rectangle.',
+        defaultMessage: 'Click and drag to draw rectangle, [esc] to cancel',
       });
     default:
       return i18n.translate('maps.drawFilterDefault.tooltipContent', {
-        defaultMessage: 'Click to start shape. Double click to finish.',
+        defaultMessage: 'Click to start shape. Double click to finish, [esc] to cancel',
       });
   }
 };
@@ -47,7 +51,7 @@ export const DrawTooltip = ({ map, mode, onCancel }: Props) => {
 
   useEffect(() => {
     // remove previous popup
-    function onMouseMoveMap(e: MapEventType['mousemove']) {
+    function onMouseMove(e: MapEventType['mousemove']) {
       map.getCanvas().style.cursor = 'crosshair'; // Changes cursor to '+'
       hoverPopupRef.current
         .setLngLat(e.lngLat)
@@ -56,11 +60,11 @@ export const DrawTooltip = ({ map, mode, onCancel }: Props) => {
         .addTo(map);
     }
 
-    function onMouseMoveOut() {
+    function onMouseOut() {
       hoverPopupRef.current.remove();
     }
 
-    function onKeyDown(e: KeyboardEvent) {
+    function onKeyUp(e: KeyboardEvent) {
       if (isEscapeKey(e)) {
         onCancel();
       }
@@ -70,18 +74,18 @@ export const DrawTooltip = ({ map, mode, onCancel }: Props) => {
       map.getCanvas().style.cursor = '';
       hoverPopupRef.current.remove();
       // remove tooltip when users mouse move over a point
-      map.off('mousemove', onMouseMoveMap);
-      map.off('mouseout', onMouseMoveOut);
-      map.getContainer().removeEventListener('keydown', onKeyDown);
+      map.off(MOUSE_MOVE_EVENT_TYPE, onMouseMove);
+      map.off(MOUSE_OUT_EVENT_TYPE, onMouseOut);
+      map.getContainer().removeEventListener(KEY_UP_EVENT_TYPE, onKeyUp);
     }
 
     if (map && mode === FILTER_DRAW_MODE.NONE) {
       resetAction();
     } else {
       // add tooltip when users mouse move over a point
-      map.on('mousemove', onMouseMoveMap);
-      map.on('mouseout', onMouseMoveOut);
-      map.getContainer().addEventListener('keydown', onKeyDown);
+      map.on(MOUSE_MOVE_EVENT_TYPE, onMouseMove);
+      map.on(MOUSE_OUT_EVENT_TYPE, onMouseOut);
+      map.getContainer().addEventListener(KEY_UP_EVENT_TYPE, onKeyUp);
     }
     return () => {
       // remove tooltip when users mouse move over a point